refactor: edit plan page refactor & show last update date
